backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
weasyprint
pydantic
starlette
python-multipart